# Livraria/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Livraria",
    "description": "Cadastro de Livrios",
    "contact": {
      "name": "Gabriel Brito Alves",
      "url": "https://www.linkedin.com/in/gabriel-brito-alves-7b8a2b188/",
      "email": "g.britoalves@yahoo.com.br"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Autors": {
      "get": {
        "tags": [
          "Autors"
        ],
        "summary": "Método para obter todos os cadastros disponíveis",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Autor"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Autors"
        ],
        "summary": "Método para publicar um cadastro",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Informações do autor para cadastrar",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        }
      }
    },
    "/api/Autors/{id}": {
      "get": {
        "tags": [
          "Autors"
        ],
        "summary": "Método para obter um valor específico",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do valor que será obtido",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Autors"
        ],
        "summary": "Método para alterar um cadastro já salvo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Informações do cadastro para alteração",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Autors"
        ],
        "summary": "Método para deletar cadastro",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros que será deletado",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Autor"
            }
          }
        }
      }
    },
    "/api/Categorias": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "summary": "Método para obter todos os cadastros disponíveis",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Categoria"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorias"
        ],
        "summary": "Método para publicar um cadastro",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Informaçoes do carastro",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        }
      }
    },
    "/api/Categorias/{id}": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "summary": "Método para obter um valor específico",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do valor que será obtido",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categorias"
        ],
        "summary": "Método para alterar um cadastro já salvo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Informações da categoria para alteração",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Categorias"
        ],
        "summary": "Método para deletar cadastro",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros que será deletado",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Autor": {
      "required": [
        "dataNascimento",
        "nacionalidade",
        "nome"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "nacionalidade": {
          "type": "string"
        },
        "dataNascimento": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Categoria": {
      "required": [
        "descricao"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descricao": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}